Fix comparison

+++ example/points/generate_points2.c
@@ -115,7 +115,7 @@ generate_points (const char *filename,
     mpiret = sc_io_write_at (file_handle, 0, &global_num_points,
                             sizeof (p4est_gloidx_t), sc_MPI_BYTE, &count);
     SC_CHECK_MPI (mpiret);
-    SC_CHECK_ABORT (count == sizeof (p4est_gloidx_t), "Write number of global"
+    SC_CHECK_ABORT (count == (int) sizeof (p4est_gloidx_t), "Write number of global"
                     "points: count mismatch");
   }
 
@@ -125,7 +125,7 @@ generate_points (const char *filename,
                                3 * local_num_points * sizeof (double),
                                sc_MPI_BYTE, &count);
   SC_CHECK_MPI (mpiret);
-  SC_CHECK_ABORT (count == 3 * local_num_points * sizeof (double),
+  SC_CHECK_ABORT (count == (int) 3 * local_num_points * sizeof (double),
                   "Write points: count mismatch");
